Split themes into separate files
