+++ src/styles/index.css
@@ -19,29 +19,6 @@
   font-weight: 400;
 }
 
-:root {
-  padding: 0px;
-  margin: 0px;
-  box-sizing: border-box !important;
-  word-break: break-word;
-  --color-instance: #1e1e20;
-  --color-instance-popover-bg: #222222;
-  --Mail-Background: rgba(49, 51, 56, 1);
-  --new-message-text: black;
-  --bg-primary: rgba(31, 32, 35, 1);
-  --bg-2: #27282c;
-  --bg-3: rgba(0, 0, 0, 0.1);
-  --unread: #4297e2;
-  --danger: #b14646;
-  --apps-circle: #1f2023;
-  --green: #5eb049;
-}
-
-body {
-  margin: 0px;
-  overflow: hidden;
-}
-
 .image-container {
   position: relative;
 }
@@ -68,6 +45,7 @@ body {
 ::-webkit-scrollbar-track {
   background-color: transparent;
 }
+
 ::-webkit-scrollbar-track:hover {
   background-color: transparent;
 }
@@ -84,6 +62,7 @@ body {
   border: 4px solid transparent;
   transition: 0.3s background-color;
 }
+
 ::-webkit-scrollbar-thumb:hover {
   background-color: #363636;
 }
@@ -130,8 +109,4 @@ body {
 html,
 body {
   overscroll-behavior: none !important;
-}
-
-.swiper {
-  width: 100%;
 }
