+++ src/styles/theme-common.ts
@@ -1,0 +1,85 @@
+// Extend the Theme interface
+const commonThemeOptions = {
+  typography: {
+    fontFamily: ['Inter'].join(','),
+    h1: {
+      fontSize: '2rem',
+      fontWeight: 600,
+    },
+    h2: {
+      fontSize: '1.75rem',
+      fontWeight: 500,
+    },
+    h3: {
+      fontSize: '1.5rem',
+      fontWeight: 500,
+    },
+    h4: {
+      fontSize: '1.25rem',
+      fontWeight: 500,
+    },
+    h5: {
+      fontSize: '1rem',
+      fontWeight: 500,
+    },
+    h6: {
+      fontSize: '0.875rem',
+      fontWeight: 500,
+    },
+    body: {
+      margin: '0px',
+      overflow: 'hidden',
+    },
+    body1: {
+      fontSize: '16px',
+      fontWeight: 400,
+      lineHeight: 1.5,
+      letterSpacing: 'normal',
+    },
+    body2: {
+      fontSize: '18px',
+      fontWeight: 400,
+      lineHeight: 1.4,
+      letterSpacing: '0.2px',
+    },
+  },
+  spacing: 8,
+  shape: {
+    borderRadius: 4,
+  },
+  breakpoints: {
+    values: {
+      xs: 0,
+      sm: 600,
+      md: 900,
+      lg: 1200,
+      xl: 1536,
+    },
+  },
+  components: {
+    MuiButton: {
+      styleOverrides: {
+        root: {
+          backgroundColor: 'inherit',
+          transition: 'filter 0.3s ease-in-out',
+          '&:hover': {
+            filter: 'brightness(1.1)',
+          },
+        },
+      },
+      defaultProps: {
+        disableElevation: true,
+        disableRipple: true,
+      },
+    },
+    MuiModal: {
+      styleOverrides: {
+        root: {
+          zIndex: 50000,
+        },
+      },
+    },
+  },
+};
+
+export { commonThemeOptions };
